ref(selection): separate @lexical/selection/index into multiple files
